\name{ShinySRAdb-package}
\alias{ShinySRAdb-package}
\alias{ShinySRAdb}
\docType{package}
\title{
\packageTitle{ShinySRAdb}
}
\description{
\packageDescription{ShinySRAdb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ShinySRAdb}
\packageIndices{ShinySRAdb}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ShinySRAdb}

Maintainer: \packageMaintainer{ShinySRAdb}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
